chore: make pages directory to split components
