+++ frontend/kick-covid-19/src/App.js
@@ -2,9 +2,9 @@ import React from "react";
 import "./assets/main.css";
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import Header from "./components/Header.js";
-import Home from "./components/Home.js";
-import Blog from "./components/Blog.js";
-import Resources from "./components/Resources.js";
+import Home from "./pages/Home";
+import Blog from "./pages/Blog.js";
+import Resources from "./pages/Resources.js";
 import Footer from "./components/Footer.js";
 
 const App = () => (
