+++ frontend/kick-covid-19/src/pages/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Card from "./Card";
+import Card from "../components/Card";
 
 function Home() {
   return (
