+++ frontend/kick-covid-19/src/pages/Resources.js
